requirements.txt
tensorflow-gpu==1.15.5
flask
loguru
jsonpickle
opencv-python
imageio==2.9.0
tf-slim==1.1.0
Pillow==8.1.0
h5py==2.10.0
imageio-ffmpeg==0.4.3